Remove more redundancy in tests
