+++ test/binary_search.test.js
@@ -1,15 +1,13 @@
-var { RabinKarp } = require('../dist/js-data-structs.cjs');
+var { BinarySearch } = require('../dist/js-data-structs.cjs');
 
 describe('Check Binary search', () => {
-    var text = 'AABAACAADAABAABA';
+    var arr = [1, 2, 3, 4, 5, 6, 7, 8, 9];
 
-    it('should return the indices of the pattern if matched', () => {
-        var pattern = 'AABA';
-        expect(RabinKarp(pattern, text, 101, 256)).toEqual([0, 9, 12]);
+    it('should return the index of the element if found', () => {
+        expect(BinarySearch(arr, 0, arr.length, 6)).toBe(5);
     });
 
-    it('should return empty array if the element is not found', () => {
-        var pattern = 'AAXBA';
-        expect(RabinKarp(pattern, text, 101, 256)).toEqual([]);
+    it('should return the -1 if the element is not found', () => {
+        expect(BinarySearch(arr, 0, arr.length, 11)).toBe(-1);
     });
 });
