+++ test/queue.test.js
@@ -1,13 +1,13 @@
 var { Queue } = require('../dist/js-data-structs.cjs');
 
 describe('Check queue functions', () => {
+    var q = Queue();
+
     it('should create an empty queue', () => {
-        var q = Queue();
         expect(q.isEmpty()).toBe(true);
     });
 
     it('should show the first element in the queue', () => {
-        var q = Queue();
         q.enqueue(4);
         q.enqueue(1);
         q.enqueue(7);
@@ -18,13 +18,6 @@ describe('Check queue functions', () => {
     });
 
     it('should return the dequeued element', () => {
-        var q = Queue();
-        q.enqueue(4);
-        q.enqueue(1);
-        q.enqueue(7);
-        q.enqueue(3);
-        q.enqueue(9);
-
         q.dequeue();
         q.dequeue();
         var x = q.dequeue();
@@ -33,13 +26,6 @@ describe('Check queue functions', () => {
     });
 
     it('should be empty', () => {
-        var q = Queue();
-        q.enqueue(4);
-        q.enqueue(1);
-        q.enqueue(7);
-        q.enqueue(3);
-
-        q.dequeue();
         q.dequeue();
         q.dequeue();
         q.dequeue();
