+++ test/stack.test.js
@@ -1,16 +1,15 @@
 var { Stack } = require('../dist/js-data-structs.cjs');
 
 describe('Check stack functions', () => {
+    var stk = Stack();
+
     it('should create an empty stack', () => {
-        var stk = Stack();
         expect(stk.isEmpty()).toBe(true);
         stk.push(1);
         expect(stk.isEmpty()).toBe(false);
     });
 
     it('should show most recently added element', () => {
-        var stk = Stack();
-        stk.push(1);
         stk.push(2);
         stk.push(3);
 
@@ -18,11 +17,6 @@ describe('Check stack functions', () => {
     });
 
     it('should return the popped element', () => {
-        var stk = Stack();
-        stk.push(1);
-        stk.push(2);
-        stk.push(3);
-
         var x = stk.pop();
         x = stk.pop();
 
@@ -30,12 +24,6 @@ describe('Check stack functions', () => {
     });
 
     it('should be empty', () => {
-        var stk = Stack();
-        stk.push(1);
-        stk.push(2);
-        stk.push(3);
-        stk.pop();
-        stk.pop();
         stk.pop();
 
         expect(stk.isEmpty()).toBe(true);
