+++ test/trie.test.js
@@ -1,9 +1,9 @@
 var { Trie, alphabetSize } = require('../dist/js-data-structs.cjs');
 
 describe('Check trie functions', () => {
-    it('should be able to add and search for words in the trie', () => {
-        var tr = Trie(alphabetSize);
+    var tr = Trie(alphabetSize);
 
+    it('should be able to add and search for words in the trie', () => {
         var list = ['the', 'a', 'there', 'answer', 'any', 'by', 'bye', 'their'];
         list.forEach((word) => {
             tr.insertWord(word);
